refactor(test): 拆解最后两个大杂烩文件 test_support_modules / test_more_support_modules
test_support_modules.py -> agent/tools/{test_message_push,test_memorize,
test_web_search,test_base}.py、memory2/test_memorizer.py、
core/common/test_timekit.py、agent/test_context.py、bus/test_queue.py、
test_repository_entrypoints.py（仓库级断言，不属于任何 backend 模块）

test_more_support_modules.py -> agent/test_provider.py、
agent/mcp/test_registry.py、proactive_v2/{test_anyaction,test_memory_sampler}.py、
infra/channels/test_group_filter.py、memory2/test_models.py、test_main.py

tests/backend 根目录下不再有横跨多模块的大杂烩文件。

Refs #194
