--- tests/backend/test_support_modules.py
@@ -1,860 +1,0 @@
-from __future__ import annotations
-from typing import Any, cast
-
-import asyncio
-import json
-from datetime import datetime, timezone
-from pathlib import Path
-from unittest.mock import AsyncMock, MagicMock
-
-import pytest
-
-from agent.context import ContextBuilder, ContextRequest
-from agent.prompting import SYSTEM_CONTEXT_FRAME_MARKER
-from agent.tools.base import Tool
-from agent.tools.memorize import MemorizeTool
-from agent.tools.message_push import MessagePushTool
-from agent.tools.web_search import WebSearchTool
-from bus.events import InboundMessage, OutboundMessage
-from bus.queue import MessageBus
-from core.common import timekit
-from core.roles import RoleStore
-from plugins.default_memory.engine import DefaultMemoryEngine
-from memory2.memorizer import Memorizer
-from memory2.store import MemoryStore2
-
-
-def _make_default_engine(
-    *,
-    retriever=None,
-    memorizer=None,
-    tagger=None,
-):
-    engine = DefaultMemoryEngine.__new__(DefaultMemoryEngine)
-    engine._config = None
-    engine._workspace = Path(".")
-    engine._provider = None
-    engine._light_provider = None
-    engine._light_model = ""
-    engine._v1_store = None
-    engine._v2_store = None
-    engine._embedder = None
-    engine._memorizer = memorizer
-    engine._retriever = retriever
-    engine._tagger = tagger
-    engine._post_response_worker = None
-    engine._event_bus = None
-    engine._consolidation = None
-    engine.closeables = []
-    return engine
-
-
-def _memorize_tool(engine) -> MemorizeTool:
-    spec = engine.tool_profile().memorize
-    assert spec is not None
-    return MemorizeTool(engine, spec)
-
-
-class _DummyTool(Tool):
-    @property
-    def name(self) -> str:
-        return "dummy"
-
-    @property
-    def description(self) -> str:
-        return "dummy description"
-
-    @property
-    def parameters(self) -> dict:
-        return {
-            "type": "object",
-            "properties": {
-                "name": {"type": "string", "minLength": 2},
-                "count": {"type": "integer", "minimum": 1, "maximum": 3},
-                "mode": {"type": "string", "enum": ["a", "b"]},
-                "items": {"type": "array", "items": {"type": "number"}},
-            },
-            "required": ["name", "count"],
-        }
-
-    async def execute(self, **kwargs) -> str:
-        return json.dumps(kwargs, ensure_ascii=False)
-
-
-@pytest.mark.asyncio
-async def test_message_push_tool_covers_success_failure_and_fallbacks():
-    tool = MessagePushTool()
-    sent = {"text": [], "stream_text": [], "file": [], "image": []}
-
-    async def text(chat_id: str, message: str) -> None:
-        sent["text"].append((chat_id, message))
-
-    async def stream_text(chat_id: str, message: str) -> None:
-        sent["stream_text"].append((chat_id, message))
-
-    async def file(chat_id: str, path: str, name: str | None) -> None:
-        sent["file"].append((chat_id, path, name))
-
-    async def image(chat_id: str, path: str) -> None:
-        sent["image"].append((chat_id, path))
-
-    tool.register_channel(
-        "telegram",
-        text=text,
-        stream_text=stream_text,
-        file=file,
-        image=image,
-    )
-    result = await tool.execute(
-        channel="telegram",
-        chat_id=123,
-        message="hello",
-        file="/tmp/demo.txt",
-        image="https://img",
-    )
-
-    assert "文本已发送" in result
-    assert "文件 'demo.txt' 已发送" in result
-    assert "图片已发送" in result
-    assert sent["text"] == []
-    assert sent["stream_text"] == [("123", "hello")]
-    assert sent["file"] == [("123", "/tmp/demo.txt", "demo.txt")]
-    assert sent["image"] == [("123", "https://img")]
-
-    assert await tool.execute(channel="telegram", chat_id=1) == (
-        "错误：message、file、image 至少提供一个"
-    )
-    assert "未注册" in await tool.execute(channel="qq", chat_id=1, message="x")
-
-    tool.register_channel("limited", text=text)
-    limited = await tool.execute(
-        channel="limited", chat_id=1, file="/tmp/a.txt", image="/tmp/a.png"
-    )
-    assert "不支持发送文件" in limited
-    assert "不支持发送图片" in limited
-
-    async def broken(chat_id: str, message: str) -> None:
-        raise RuntimeError("send failed")
-
-    tool.register_channel("broken", text=broken)
-    assert "发送失败" in await tool.execute(channel="broken", chat_id=1, message="x")
-
-
-@pytest.mark.asyncio
-async def test_memorize_tool_cover_branches(
-    monkeypatch: pytest.MonkeyPatch, tmp_path: Path
-):
-    memorizer = MagicMock()
-    memorizer.save_item_with_supersede = AsyncMock(return_value="new:mem-1")
-
-    class _Tagger:
-        async def tag(self, summary: str) -> dict[str, str]:
-            assert summary == "记住这条流程"
-            return {"scope": "task"}
-
-    tool = _memorize_tool(
-        _make_default_engine(
-            retriever=MagicMock(),
-            memorizer=memorizer,
-            tagger=cast(Any, _Tagger()),
-        )
-    )
-    result = await tool.execute(
-        summary="记住这条流程",
-        memory_kind="procedure",
-        steps=["先查", "再做"],
-        role_id="mira",
-    )
-
-    payload = json.loads(result)
-    assert payload["item_id"] == "mem-1"
-    assert payload["status"] == "new"
-    extra = memorizer.save_item_with_supersede.await_args.kwargs["extra"]
-    assert extra["trigger_tags"] == {"scope": "task"}
-    assert extra["rule_schema"]["required_tools"] == []
-    assert extra["rule_schema"]["forbidden_tools"] == []
-
-    class _BadTagger:
-        async def tag(self, summary: str) -> dict[str, str]:
-            raise RuntimeError("bad")
-
-    bad = _memorize_tool(
-        _make_default_engine(
-            retriever=MagicMock(),
-            memorizer=memorizer,
-            tagger=cast(Any, _BadTagger()),
-        )
-    )
-    await bad.execute(summary="普通偏好", memory_kind="procedure", role_id="mira")
-    await bad.execute(summary="偏好", memory_kind="preference", role_id="mira")
-
-
-@pytest.mark.asyncio
-async def test_memorize_tool_should_not_create_second_active_procedure_when_incremental_update():
-    class _Embedder:
-        async def embed(self, text: str) -> list[float]:
-            return [1.0, 0.0]
-
-    store = MemoryStore2(":memory:")
-    memorizer = Memorizer(store, cast(Any, _Embedder()))
-    tool = _memorize_tool(
-        _make_default_engine(
-            retriever=MagicMock(),
-            memorizer=memorizer,
-        )
-    )
-
-    await memorizer.save_item(
-        summary="查询 Steam 游戏信息时，必须先使用 steam_mcp 工具查询游戏详情，再用 web_search 补充验证价格和评价信息。",
-        memory_type="procedure",
-        extra={
-            "steps": [
-                "使用 steam_mcp 工具查询游戏详情",
-                "使用 web_search 补充验证价格和评价",
-            ],
-            "tool_requirement": "steam_mcp",
-            "role_id": "mira",
-        },
-        source_ref="seed",
-    )
-
-    await tool.execute(
-        summary="查询 Steam 游戏信息时，先判断区服（大陆区/港区/美区），再使用 steam_mcp 工具查询游戏详情。",
-        memory_kind="procedure",
-        tool_requirement="steam_mcp",
-        steps=["判断目标区服", "使用 steam_mcp 工具查询游戏详情"],
-        role_id="mira",
-    )
-
-    rows = store._db.execute(
-        "SELECT id, summary FROM memory_items WHERE memory_type='procedure' AND status='active'"
-    ).fetchall()
-    assert len(rows) == 1
-    assert "steam_mcp" in rows[0][1]
-    assert "区服" in rows[0][1]
-
-
-@pytest.mark.asyncio
-async def test_memorizer_profile_supersede_keeps_high_emotional_weight_item_under_092():
-    class _Embedder:
-        async def embed(self, text: str) -> list[float]:
-            mapping = {
-                "用户仍在等待 offer": [1.0, 0.0],
-                "用户开始等待新的 offer": [0.91, 0.4146],
-            }
-            return mapping[text]
-
-    store = MemoryStore2(":memory:")
-    memorizer = Memorizer(store, cast(Any, _Embedder()))
-
-    await memorizer.save_item(
-        summary="用户仍在等待 offer",
-        memory_type="profile",
-        extra={"category": "status"},
-        source_ref="old",
-        emotional_weight=8,
-    )
-    await memorizer.save_item_with_supersede(
-        summary="用户开始等待新的 offer",
-        memory_type="profile",
-        extra={"category": "status"},
-        source_ref="new",
-    )
-
-    rows = store._db.execute(
-        "SELECT source_ref, status FROM memory_items WHERE memory_type='profile' ORDER BY source_ref"
-    ).fetchall()
-    assert rows == [("new", "active"), ("old", "active")]
-
-
-@pytest.mark.asyncio
-async def test_memorizer_profile_supersede_retires_low_emotional_weight_item_at_091():
-    class _Embedder:
-        async def embed(self, text: str) -> list[float]:
-            mapping = {
-                "用户仍在等待 offer": [1.0, 0.0],
-                "用户开始等待新的 offer": [0.91, 0.4146],
-            }
-            return mapping[text]
-
-    store = MemoryStore2(":memory:")
-    memorizer = Memorizer(store, cast(Any, _Embedder()))
-
-    await memorizer.save_item(
-        summary="用户仍在等待 offer",
-        memory_type="profile",
-        extra={"category": "status"},
-        source_ref="old",
-        emotional_weight=0,
-    )
-    await memorizer.save_item_with_supersede(
-        summary="用户开始等待新的 offer",
-        memory_type="profile",
-        extra={"category": "status"},
-        source_ref="new",
-    )
-
-    rows = store._db.execute(
-        "SELECT source_ref, status FROM memory_items WHERE memory_type='profile' ORDER BY source_ref"
-    ).fetchall()
-    assert rows == [("new", "active"), ("old", "superseded")]
-
-
-@pytest.mark.asyncio
-async def test_memorize_tool_should_coerce_language_reply_rule_to_preference():
-    memorizer = MagicMock()
-    memorizer.save_item_with_supersede = AsyncMock(return_value="new:mem-1")
-    tool = _memorize_tool(
-        _make_default_engine(
-            retriever=MagicMock(),
-            memorizer=memorizer,
-        )
-    )
-
-    await tool.execute(
-        summary="之后跟我说话只用中文，不要夹杂英文，专有名词也尽量翻译。",
-        memory_kind="procedure",
-        role_id="mira",
-    )
-
-    assert (
-        memorizer.save_item_with_supersede.await_args.kwargs["memory_type"]
-        == "preference"
-    )
-
-
-@pytest.mark.asyncio
-async def test_web_search_covers_filters(monkeypatch: pytest.MonkeyPatch):
-    class _Response:
-        def __init__(self, text: str) -> None:
-            self.text = text
-
-        def raise_for_status(self) -> None:
-            return None
-
-    class _Client:
-        def __init__(self, timeout: float) -> None:
-            self.timeout = timeout
-
-        async def __aenter__(self):
-            return self
-
-        async def __aexit__(self, exc_type, exc, tb):
-            return False
-
-        async def post(self, url: str, json: dict, headers: dict) -> _Response:
-            assert json["params"]["arguments"]["numResults"] == 20
-            assert json["params"]["arguments"]["livecrawl"] == "preferred"
-            assert json["params"]["arguments"]["type"] == "deep"
-            return _Response(
-                'data: {"result":{"content":[{"text":"hello world"}]}}\n\n'
-            )
-
-    monkeypatch.setattr("httpx.AsyncClient", _Client)
-    result = json.loads(
-        await WebSearchTool().execute(
-            query="搜索 网络",
-            num_results=99,
-            livecrawl="preferred",
-            type="deep",
-        )
-    )
-    assert result["result"] == "hello world"
-
-    class _BadClient(_Client):
-        async def post(self, url: str, json: dict, headers: dict) -> _Response:
-            raise RuntimeError("net down")
-
-    monkeypatch.setattr("httpx.AsyncClient", _BadClient)
-    result = json.loads(await WebSearchTool().execute(query="x"))
-    assert "搜索失败" in result["error"]
-
-    class _EmptyClient(_Client):
-        async def post(self, url: str, json: dict, headers: dict) -> _Response:
-            return _Response("data: not-json\n\ndata: {}")
-
-    monkeypatch.setattr("httpx.AsyncClient", _EmptyClient)
-    result = json.loads(await WebSearchTool().execute(query="x"))
-    assert result["count"] == 0
-
-
-def test_tool_base_and_timekit_cover_branches(
-    monkeypatch: pytest.MonkeyPatch, tmp_path: Path
-):
-    tool = _DummyTool()
-    errors = tool.validate_params(
-        {"name": "x", "count": 5, "mode": "c", "items": ["a"]}
-    )
-    assert "name 最短 2 个字符" in errors
-    assert "count 须 <= 3" in errors
-    assert "mode 须为以下值之一" in errors[2]
-    assert "[0] 应为 number 类型" in errors[3]
-    assert tool.validate_params({})[:2] == ["缺少必填字段：name", "缺少必填字段：count"]
-    assert tool.to_schema()["function"]["name"] == "dummy"
-
-    class _BadSchemaTool(_DummyTool):
-        @property
-        def parameters(self) -> dict:
-            return {"type": "array"}
-
-    with pytest.raises(ValueError):
-        _BadSchemaTool().validate_params({})
-
-    with pytest.raises(TypeError, match="必须定义字段：description, parameters"):
-
-        class _MissingTool(Tool):
-            name = "bad"
-
-            async def execute(self, **kwargs) -> str:
-                return "ok"
-
-    with pytest.raises(TypeError, match="字段不能为空：name, description, parameters"):
-
-        class _EmptyTool(Tool):
-            name = ""
-            description = ""
-            parameters = {}
-
-            async def execute(self, **kwargs) -> str:
-                return "ok"
-
-    parsed = timekit.parse_iso("2025-06-01T09:00:00Z")
-    assert parsed and parsed.tzinfo is not None
-    assert timekit.parse_iso("bad") is None
-    assert timekit.format_iso(datetime(2025, 1, 1)).endswith("+00:00")
-    logger = MagicMock()
-    assert str(timekit.safe_zone("bad/zone", logger=logger)) == "UTC"
-    logger.warning.assert_called_once()
-    assert timekit.local_now("UTC").tzinfo is not None
-    assert timekit.utcnow().tzinfo is not None
-
-
-def test_context_builder_builds_prompt_messages_and_assistant_blocks(
-    monkeypatch: pytest.MonkeyPatch, tmp_path: Path
-):
-    class _Skills:
-        def __init__(self, workspace: Path) -> None:
-            self.workspace = workspace
-
-        def get_always_skills(self) -> list[str]:
-            return ["always"]
-
-        def load_skills_for_context(self, names: list[str]) -> str:
-            return ",".join(names)
-
-        def build_skills_summary(self) -> str:
-            return "skill summary"
-
-    class _Memory:
-        def read_profile(self) -> str:
-            return "memory block"
-
-        def read_self(self) -> str:
-            return "self note"
-
-        def read_recent_context(self) -> str:
-            return ""
-
-        def get_memory_context(self) -> str:
-            return "memory block"
-
-    monkeypatch.setattr("agent.context.SkillsLoader", _Skills)
-    monkeypatch.setattr(
-        "agent.context.build_agent_static_identity_prompt", lambda **_: "identity"
-    )
-    monkeypatch.setattr(
-        "agent.context.build_telegram_rendering_prompt", lambda: "\ntelegram prompt"
-    )
-    monkeypatch.setattr(
-        "agent.context.build_skills_catalog_prompt", lambda text: f"catalog:{text}"
-    )
-
-    image = tmp_path / "a.png"
-    image.write_bytes(b"\x89PNG\r\n\x1a\n")
-    now = datetime.now(timezone.utc)
-    role_store = RoleStore(tmp_path)
-    role_store.create_role(
-        role_id="mira",
-        name="Mira",
-        description="desktop role",
-        system_prompt="你现在要用更温柔的风格说话。",
-        background="来自深海城的向导。",
-        runtime_config={"shared_memory_enabled": True, "model": "deepseek-chat"},
-    )
-    role_metadata = {"role_id": "mira"}
-
-    builder = ContextBuilder(tmp_path, _Memory())  # type: ignore[arg-type]
-    result = builder.render(
-        ContextRequest(
-            history=[],
-            current_message="",
-            skill_names=["extra"],
-            message_timestamp=now,
-            retrieved_memory_block="retrieved",
-        ),
-        session_metadata=role_metadata,
-    )
-    prompt = result.system_prompt
-    context_frame = result.messages[-2]["content"]
-    assert "identity" in prompt
-    assert "## 行为规范" in prompt
-    assert "retrieved" not in prompt
-    assert context_frame.startswith(SYSTEM_CONTEXT_FRAME_MARKER)
-    assert "retrieved" in context_frame
-    assert "memory block" in prompt
-    assert "角色自我认知" in prompt
-    assert "## 环境" in prompt
-    assert "# Memes" not in prompt
-    assert "<meme:shy>" not in prompt
-    assert "catalog:skill summary" in prompt
-    assert [item.name for item in builder.last_debug_breakdown][:3] == [
-        "role_cache_prefix",
-        "active_role",
-        "identity",
-    ]
-
-    result2 = builder.render(
-        ContextRequest(
-            history=[],
-            current_message="",
-            skill_names=["extra"],
-            message_timestamp=now,
-            retrieved_memory_block="retrieved",
-        ),
-        session_metadata=role_metadata,
-    )
-    assert result2.system_prompt
-    identity_meta = next(
-        item for item in builder.last_debug_breakdown if item.name == "identity"
-    )
-    assert identity_meta.cache_hit is True
-
-    messages = builder.render(
-        ContextRequest(
-            history=[{"role": "assistant", "content": "hi"}],
-            current_message="hello",
-            media=["https://img", str(image), str(tmp_path / "bad.txt")],
-            skill_names=["extra"],
-            channel="telegram",
-            chat_id="42",
-        ),
-        session_metadata=role_metadata,
-    ).messages
-    assert messages[0]["role"] == "system"
-    assert "## 环境" in messages[0]["content"]
-    assert "## Current Session" in messages[0]["content"]
-    assert messages[-1]["role"] == "user"
-    assert len(messages[-1]["content"]) == 3
-    stamped_message = messages[-1]["content"][-1]["text"]
-    assert stamped_message.startswith("[当前消息时间:")
-    assert "request_time=" in stamped_message
-    assert "今天=" in stamped_message
-    assert "昨天=" in stamped_message
-    assert "明天=" in stamped_message
-    assert "后天=" in stamped_message
-    assert "weekday=" in stamped_message
-    assert builder.last_assembled_contexts["turn_injection_context"] == {}
-
-    turn_injection = builder.build_turn_injection_context(turn_injection_prompt="pref")
-    render_result = builder.render(
-        ContextRequest(
-            history=[{"role": "assistant", "content": "hi"}],
-            current_message="hello",
-            media=["https://img", str(image), str(tmp_path / "bad.txt")],
-            skill_names=["extra"],
-            channel="telegram",
-            chat_id="42",
-            message_timestamp=now,
-            turn_injection_prompt="pref",
-        ),
-        session_metadata=role_metadata,
-    )
-    assert render_result.system_prompt
-    assert render_result.turn_injection_context == turn_injection
-    assert render_result.messages
-    assert render_result.messages[-2]["role"] == "user"
-    assert render_result.messages[-2]["content"].startswith(SYSTEM_CONTEXT_FRAME_MARKER)
-    assert "pref" in render_result.messages[-2]["content"]
-
-    custom_telegram = builder.render(
-        ContextRequest(
-            history=[],
-            current_message="hello",
-            channel="telegram_work",
-            chat_id="42",
-            message_timestamp=now,
-        ),
-        session_metadata=role_metadata,
-    )
-    assert "telegram prompt" in custom_telegram.messages[0]["content"]
-
-    qqbot = builder.render(
-        ContextRequest(
-            history=[],
-            current_message="hello",
-            channel="qqbot",
-            chat_id="c2c:user-1",
-            message_timestamp=now,
-        ),
-        session_metadata=role_metadata,
-    )
-    assert "## 官方 QQBot 渠道规则（硬性）" in qqbot.messages[0]["content"]
-    assert "必须使用 `message_push` 的 `channel=qqbot`" in qqbot.messages[0]["content"]
-    assert "不得把官方 QQBot 写成 `channel=qq`" in qqbot.messages[0]["content"]
-    assert "c2c:<user_openid>" in qqbot.messages[0]["content"]
-
-    media_only_messages = builder.render(
-        ContextRequest(
-            history=[],
-            current_message="",
-            media=["https://img"],
-            skill_names=["extra"],
-            message_timestamp=now,
-        ),
-        session_metadata=role_metadata,
-    ).messages
-    media_only_text = media_only_messages[-1]["content"][-1]["text"]
-    assert media_only_text.startswith("[当前消息时间:")
-    assert "request_time=" in media_only_text
-    assert "今天=" in media_only_text
-
-    text_media_builder = ContextBuilder(
-        tmp_path,
-        _Memory(),  # type: ignore[arg-type]
-        multimodal=False,
-    )
-    text_media_messages = text_media_builder.render(
-        ContextRequest(
-            history=[],
-            current_message="看看这张图",
-            media=[str(image)],
-            skill_names=["extra"],
-            message_timestamp=now,
-        ),
-        session_metadata=role_metadata,
-    ).messages
-    text_media_content = text_media_messages[-1]["content"]
-    assert isinstance(text_media_content, str)
-    assert str(image) in text_media_content
-    assert "当前模型不支持多模态，无法处理图片内容。" in text_media_content
-    assert "image_url" not in text_media_content
-
-    note_path = tmp_path / "chat-note.md"
-    note_path.write_text("# note\n\nhello\n", encoding="utf-8")
-    text_attachment_messages = builder.render(
-        ContextRequest(
-            history=[],
-            current_message="这是附件",
-            media=[str(note_path)],
-            skill_names=["extra"],
-            message_timestamp=now,
-        ),
-        session_metadata=role_metadata,
-    ).messages
-    text_attachment_content = text_attachment_messages[-1]["content"]
-    assert isinstance(text_attachment_content, str)
-    assert "[附加文件]" in text_attachment_content
-    assert str(note_path) in text_attachment_content
-    assert "read_file(path=" in text_attachment_content
-
-    mixed_media_messages = builder.render(
-        ContextRequest(
-            history=[],
-            current_message="图和文档都在这",
-            media=[str(image), str(note_path)],
-            skill_names=["extra"],
-            message_timestamp=now,
-        ),
-        session_metadata=role_metadata,
-    ).messages
-    mixed_media_content = mixed_media_messages[-1]["content"]
-    assert isinstance(mixed_media_content, list)
-    assert mixed_media_content[-1]["type"] == "text"
-    assert str(note_path) in mixed_media_content[-1]["text"]
-    assert "read_file(path=" in mixed_media_content[-1]["text"]
-
-    role_memory_root = tmp_path / "roles" / "mira" / "memory"
-    role_memory_root.mkdir(parents=True, exist_ok=True)
-    (role_memory_root / "SELF.md").write_text("# 角色背景\n\n来自深海城的向导。\n", encoding="utf-8")
-    (role_memory_root / "MEMORY.md").write_text("# 关系基线\n\n来源: seed:first_impression\n", encoding="utf-8")
-    role_messages = builder.render(
-        ContextRequest(
-            history=[],
-            current_message="hello",
-            skill_names=["extra"],
-            message_timestamp=now,
-        ),
-        session_metadata={"role_id": "mira"},
-    ).messages
-    role_prompt = role_messages[0]["content"]
-    assert "role_id=mira" in role_prompt
-    assert "[role_background]" in role_prompt
-    assert "[role_runtime_config]" in role_prompt
-    assert "## Active Role: Mira" in role_prompt
-    assert "你现在要用更温柔的风格说话。" in role_prompt
-    assert "你是一个用户创建的角色" not in role_prompt
-    assert "tool_search" in role_prompt
-
-    role_prompt_cross_channel = builder.render(
-        ContextRequest(
-            history=[],
-            current_message="hello",
-            skill_names=["extra"],
-            channel="telegram",
-            chat_id="42",
-            message_timestamp=now,
-        ),
-        session_metadata={"role_id": "mira"},
-    ).messages[0]["content"]
-    assert "role_id=mira" in role_prompt_cross_channel
-    assert "[role_background]" in role_prompt_cross_channel
-    role_prefix = role_prompt.split("## Active Role: Mira", 1)[0]
-    role_prefix_cross_channel = role_prompt_cross_channel.split("## Active Role: Mira", 1)[0]
-    assert role_prefix == role_prefix_cross_channel
-
-
-def test_context_builder_reproduces_temporal_conflict_baseline(
-    monkeypatch: pytest.MonkeyPatch, tmp_path: Path
-):
-    class _Skills:
-        def __init__(self, workspace: Path) -> None:
-            self.workspace = workspace
-
-        def get_always_skills(self) -> list[str]:
-            return []
-
-        def load_skills_for_context(self, names: list[str]) -> str:
-            return ""
-
-        def build_skills_summary(self) -> str:
-            return ""
-
-    class _Memory:
-        def read_profile(self) -> str:
-            return ""
-
-        def read_self(self) -> str:
-            return ""
-
-        def read_recent_context(self) -> str:
-            return ""
-
-        def get_memory_context(self) -> str:
-            return ""
-
-    monkeypatch.setattr("agent.context.SkillsLoader", _Skills)
-    monkeypatch.setattr(
-        "agent.context.build_agent_static_identity_prompt", lambda **_: "identity"
-    )
-    monkeypatch.setattr("agent.context.build_telegram_rendering_prompt", lambda: "")
-    monkeypatch.setattr("agent.context.build_skills_catalog_prompt", lambda text: text)
-
-    (tmp_path / "memes").mkdir()
-    (tmp_path / "memes" / "manifest.json").write_text(
-        '{"version":1,"categories":{}}',
-        encoding="utf-8",
-    )
-    RoleStore(tmp_path).create_role(
-        role_id="mira",
-        name="Mira",
-        system_prompt="你是 Mira。",
-    )
-
-    builder = ContextBuilder(tmp_path, _Memory())  # type: ignore[arg-type]
-    request_time = datetime.fromisoformat("2026-04-08T17:57:00+08:00")
-    retrieved_memory_block = """
-[item_5a9c8d59f77c] [2026-03-29 12:44] 用户表示明天下午三点有面试，因当前感到疲惫想小睡，但担心此举会打乱明天的生物钟。
-证据: 用户消息「明天我下午三点面试 我现在睡一会会打乱明天发生物钟吗有点疲惫」
-
-[item_87aa0364de9e] [2026-03-29 14:42] 用户因午睡未成功，转为练习力扣题目以准备次日下午三点的字节跳动面试。
-证据: 用户消息「没睡着做会力扣准备明天面试了」
-
-[item_recent_interview] [2026-04-07 23:10] 用户提到 4 月 9 日（周四）下午 3 点的面试安排。
-证据: 可回源原文「4 月 9 日（周四）下午 3 点」
-""".strip()
-
-    result = builder.render(
-        ContextRequest(
-            history=[],
-            current_message="你还记得明天什么时候面试吗",
-            channel="telegram",
-            chat_id="7674283004",
-            message_timestamp=request_time,
-            retrieved_memory_block=retrieved_memory_block,
-        ),
-        session_metadata={"role_id": "mira"},
-    )
-
-    system_prompt = result.messages[0]["content"]
-    context_frame = result.messages[-2]["content"]
-    user_message = result.messages[-1]["content"]
-
-    assert "request_time=2026-04-08T17:57:00+08:00" not in system_prompt
-    assert "local_date=2026-04-08" not in system_prompt
-    assert "今天=2026-04-08" not in system_prompt
-    assert "明天=2026-04-09" not in system_prompt
-    assert context_frame.startswith(SYSTEM_CONTEXT_FRAME_MARKER)
-    assert "用户表示明天下午三点有面试" in context_frame
-    assert "准备次日下午三点的字节跳动面试" in context_frame
-    assert "4 月 9 日（周四）下午 3 点" in context_frame
-    assert user_message.startswith("[当前消息时间: 2026-04-08 17:57:00")
-    assert "request_time=2026-04-08T17:57:00+08:00" in user_message
-    assert "今天=2026-04-08" in user_message
-    assert "昨天=2026-04-07" in user_message
-    assert "明天=2026-04-09" in user_message
-    assert "后天=2026-04-10" in user_message
-    assert "weekday=Wednesday" in user_message
-    assert "相对时间以此为准" in user_message
-    assert user_message.endswith("你还记得明天什么时候面试吗")
-
-
-@pytest.mark.asyncio
-async def test_message_bus_covers_flows(
-    monkeypatch: pytest.MonkeyPatch, tmp_path: Path
-):
-    bus = MessageBus()
-    await bus.publish_inbound(InboundMessage("telegram", "u", "1", "hello"))
-    inbound = await bus.consume_inbound()
-    assert inbound.session_key == "telegram:1"
-
-    sent: list[str] = []
-    attempts = {"count": 0}
-
-    async def callback(msg: OutboundMessage) -> None:
-        attempts["count"] += 1
-        if attempts["count"] == 1:
-            raise RuntimeError("first")
-        sent.append(msg.content)
-
-    bus.subscribe_outbound("telegram", callback)
-    task = asyncio.create_task(bus.dispatch_outbound())
-    await bus.publish_outbound(OutboundMessage("telegram", "1", "payload"))
-    for _ in range(300):
-        if sent:
-            break
-        await asyncio.sleep(0.01)
-    bus.stop()
-    task.cancel()
-    with pytest.raises(asyncio.CancelledError):
-        await task
-    assert sent == ["payload"]
-    assert bus.inbound_size == 0
-    assert bus.outbound_size == 0
-
-
-def test_repository_entrypoints_are_desktop_first():
-    repo_root = Path(__file__).resolve().parents[2]
-    package_json = json.loads((repo_root / "package.json").read_text(encoding="utf-8"))
-    scripts = package_json["scripts"]
-
-    assert scripts["start"] == "pnpm run desktop:start"
-    assert scripts["dev"] == "pnpm run desktop:dev"
-    assert "dashboard:dev" not in scripts
-    assert "dashboard:build" not in scripts
-    assert "build:dashboard" not in scripts
-    assert "build:dashboard:watch" not in scripts
-
-    readme = (repo_root / "README.md").read_text(encoding="utf-8")
-    assert "npm run dashboard:dev" not in readme
-    assert "npm run dashboard:build" not in readme
-    assert "main.py dashboard" not in readme
